Rename InFilter to FilterSummary
